requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.8.2
pydantic-settings==2.4.0
alembic==1.13.2
python-dateutil==2.8.2
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1
